% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{my_palette3}
\alias{my_palette3}
\title{My Custom Color Palette 3}
\format{
A named character vector.
}
\usage{
my_palette3(n)
}
\description{
A vector of colors including green, sand, and sunset tones.
}
\examples{
print(my_palette3)
}
